feat: fetch templates from github
