+++ src/interfaces/store/templates.interface.ts
@@ -1,3 +1,6 @@
+import { TemplateStorageService } from "@services";
+import { TemplateCategory, TemplateMetadata } from "@src/types/components";
+
 export interface RemoteTemplateMetadata {
 	assetDirectory: string;
 	title: string;
@@ -17,4 +20,28 @@ export interface RemoteTemplateCardWithFiles {
 export interface ProcessedRemoteCategory {
 	name: string;
 	cards: RemoteTemplateCardWithFiles[];
+}
+
+export interface TemplateState {
+	templateMap: Record<string, TemplateMetadata>;
+	isLoading: boolean;
+	sortedCategories?: TemplateCategory[];
+	error: string | null;
+	lastCommitDate?: string;
+	templateStorage: TemplateStorageService;
+
+	fetchTemplates: () => Promise<void>;
+	findTemplateByAssetDirectory: (assetDirectory: string) => TemplateMetadata | undefined;
+	getTemplateFiles: (assetDirectory: string) => Promise<Record<string, string>>;
+}
+
+export interface GitHubCommit {
+	sha: string;
+	commit: {
+		author: {
+			date: string;
+			name: string;
+		};
+		message: string;
+	};
 }
